Refactoring as discussed in PR #75 .
