+++ src/main/java/bitflow4j/Pipeline.java
@@ -7,6 +7,7 @@ import bitflow4j.io.marshall.CsvMarshaller;
 import bitflow4j.misc.TreeFormatter;
 import bitflow4j.steps.BatchHandler;
 import bitflow4j.steps.BatchPipelineStep;
+import bitflow4j.steps.MergeBatchPipelineStep;
 import bitflow4j.task.ParallelTask;
 import bitflow4j.task.StoppableTask;
 import bitflow4j.task.Task;
@@ -73,8 +74,11 @@ public class Pipeline implements TreeFormatter.FormattedNode {
         return this;
     }
 
-    public Pipeline batch(String separationTag, long timeout, boolean mapMode, BatchHandler... steps) {
-        return step(new BatchPipelineStep(separationTag, timeout, mapMode, steps));
+    public Pipeline batch(String separationTag, long timeout, boolean mergeMode, BatchHandler... steps) {
+        if (mergeMode) {
+            return step(new MergeBatchPipelineStep(separationTag, timeout, steps));
+        }
+        return step(new BatchPipelineStep(separationTag, timeout, steps));
     }
 
     public Pipeline batch(String separationTag, long timeout, BatchHandler... steps) {
