+++ src/main/java/bitflow4j/steps/BatchPipelineStep.java
@@ -1,20 +1,14 @@
 package bitflow4j.steps;
 
-import bitflow4j.AbstractPipelineStep;
 import bitflow4j.Sample;
-import bitflow4j.misc.TreeFormatter;
 import bitflow4j.script.registry.BitflowConstructor;
 import bitflow4j.script.registry.Optional;
-import bitflow4j.script.registry.RegisteredParameter;
-import bitflow4j.script.registry.RegisteredParameterList;
 import bitflow4j.task.LoopTask;
 import bitflow4j.task.TaskPool;
 
 import java.io.IOException;
-import java.util.*;
-import java.util.concurrent.ConcurrentHashMap;
+import java.util.Date;
 import java.util.logging.Level;
-import java.util.logging.Logger;
 
 /**
  * Instead of immediately handling every Sample, fill up a list of samples based on a number of criteria (time, tag values, ...).
@@ -26,97 +20,30 @@ import java.util.logging.Logger;
  * <p>
  * Created by anton on 5/8/16.
  */
-public final class BatchPipelineStep extends AbstractPipelineStep implements TreeFormatter.FormattedNode {
-
-    protected static final Logger logger = Logger.getLogger(BatchPipelineStep.class.getName());
-
-    private final String batchSeparationTag;
+public final class BatchPipelineStep extends AbstractBatchPipelineStep {
     private boolean warnedMissingSeparationTag = false;
     private String previousSeparationTagValue = null;
+    final String batchSeparationTag;
+    final long timeoutMs;
+    long startTime;
 
-    private final long timeoutMs;
-    private long startTime;
-
-    private List<Sample> window = new ArrayList<>();
-    private final List<BatchHandler> handlers = new ArrayList<>();
-
-    private Map<String, Long> timeoutMap = new ConcurrentHashMap<>();
-    private Map<String, List<Sample>> samplesForTag = new ConcurrentHashMap<>();
-    private List<String> tagList = new ArrayList<>();
-    private final boolean mapMode;
-
-    public BatchPipelineStep() {
-        this((String) null);
-    }
-
-    public BatchPipelineStep(BatchHandler... handlers) {
-        this(null, handlers);
-    }
-
-    public BatchPipelineStep(String batchSeparationTag) {
-        this(batchSeparationTag, 0);
-    }
-
-    public BatchPipelineStep(String batchSeparationTag, long timeoutMs) {
-        this(batchSeparationTag, timeoutMs, false);
+    @BitflowConstructor
+    public BatchPipelineStep(@Optional String batchSeparationTag, @Optional long timeoutMs, BatchHandler... handlers) {
+        super(handlers);
+        this.batchSeparationTag = batchSeparationTag;
+        this.timeoutMs = timeoutMs;
     }
 
     public BatchPipelineStep(String batchSeparationTag, BatchHandler... handlers) {
-        this(batchSeparationTag, 0, false, handlers);
+        this(batchSeparationTag, 0, handlers);
     }
 
-    public BatchPipelineStep(String batchSeparationTag, long timeoutMs, BatchHandler... handlers) {
-        this(batchSeparationTag, timeoutMs, false, handlers);
-    }
-
-    @BitflowConstructor
-    public BatchPipelineStep(@Optional String batchSeparationTag, @Optional long timeoutMs, @Optional boolean mapMode) {
-        this.batchSeparationTag = batchSeparationTag;
-        this.timeoutMs = timeoutMs;
-        this.mapMode = mapMode;
-    }
-
-    public BatchPipelineStep(String batchSeparationTag, long timeoutMs, boolean mapMode, BatchHandler... handlers) {
-        this.batchSeparationTag = batchSeparationTag;
-        this.timeoutMs = timeoutMs;
-        this.mapMode = mapMode;
-        this.handlers.addAll(Arrays.asList(handlers));
-    }
-
-    public static final RegisteredParameterList BATCH_STEP_PARAMETERS = new RegisteredParameterList(
-            new RegisteredParameter[]{
-                    new RegisteredParameter("separationTag", RegisteredParameter.ContainerType.Primitive, String.class, ""),
-                    new RegisteredParameter("timeout", RegisteredParameter.ContainerType.Primitive, Long.class, 0L),
-                    new RegisteredParameter("mapMode", RegisteredParameter.ContainerType.Primitive, Boolean.class, false)
-            });
-
-    public static BatchPipelineStep createFromParameters(Map<String, Object> params) throws IllegalArgumentException {
-        String separationTag = null;
-        if (params.containsKey("separationTag")) {
-            separationTag = (String) params.get("separationTag");
-        }
-        long timeout = 0;
-        if (params.containsKey("timeout")) {
-            timeout = (Long) params.get("timeout");
-        }
-        boolean mapMode = false;
-        if (params.containsKey("mapMode")) {
-            mapMode = (Boolean) params.get("mapMode");
-        }
-        return new BatchPipelineStep(separationTag, timeout, mapMode);
-    }
-
-    public void addBatchHandler(BatchHandler handler) {
-        handlers.add(handler);
+    public BatchPipelineStep(BatchHandler... handlers) {
+        this((String) null, handlers);
     }
 
     @Override
-    public void start(TaskPool pool) throws IOException {
-        super.start(pool);
-
-        // TODO refactor for allowing additional flushing modes:
-        // window size, timeout (wall clock, sample timestamps), tag change. Micro batching ("jumping" window) vs moving window.
-
+    public void threadIteration(TaskPool pool) throws IOException {
         if (timeoutMs > 0) {
             startTime = new Date().getTime();
             pool.start(new LoopTask() {
@@ -128,21 +55,7 @@ public final class BatchPipelineStep extends AbstractPipelineStep implements Tre
                 @Override
                 protected boolean executeIteration() throws IOException {
                     long currentTime = System.currentTimeMillis();
-                    if (mapMode) {
-                        flushMapResults(currentTime);
-                    } else {
-                        if (currentTime - startTime > timeoutMs) {
-                            try {
-                                boolean flushed = flushResults();
-                                if (flushed) {
-                                    logger.log(Level.INFO, String.format("Flushed batch due to timeout (%sms) for step: %s.", timeoutMs, BatchPipelineStep.this.toString()));
-                                }
-                            } catch (IOException ex) {
-                                logger.log(Level.SEVERE, "Failed to automatically flush batch", ex);
-                            }
-                        }
-                    }
-
+                    checkForFlush(currentTime);
                     return pool.sleep(timeoutMs / 2);
                 }
             });
@@ -150,29 +63,32 @@ public final class BatchPipelineStep extends AbstractPipelineStep implements Tre
     }
 
     @Override
-    public final synchronized void writeSample(Sample sample) throws IOException {
-        if (mapMode) {
-            addSampleToMaps(sample);
-        } else {
-            startTime = new Date().getTime();
-            if (shouldFlush(sample)) {
-                flushResults();
+    public void checkForFlush(long currentTime) throws IOException {
+        if (currentTime - startTime > timeoutMs) {
+            try {
+                boolean flushed = flushResults();
+                if (flushed) {
+                    logger.log(Level.INFO, String.format("Flushed batch due to timeout (%sms) for step: %s.", timeoutMs, BatchPipelineStep.this.toString()));
+                }
+            } catch (IOException ex) {
+                logger.log(Level.SEVERE, "Failed to automatically flush batch", ex);
             }
-            window.add(sample);
         }
     }
 
-    private void addSampleToMaps(Sample sample) {
-        if (timeoutMap.get(sample.getTag(batchSeparationTag)) != null) {
-            samplesForTag.get(sample.getTag(batchSeparationTag)).add(sample);
-        } else {
-            //Put Sample into timeoutMap and create new ArrayList with Sample
-            timeoutMap.put(sample.getTag(batchSeparationTag), System.currentTimeMillis() + timeoutMs);
-            List<Sample> newList = new ArrayList<>();
-            newList.add(sample);
-            samplesForTag.put(sample.getTag(batchSeparationTag), newList);
-            tagList.add(sample.getTag(batchSeparationTag));
+    @Override
+    public void addSample(Sample sample) throws IOException {
+        startTime = new Date().getTime();
+        if (shouldFlush(sample)) {
+            flushResults();
         }
+        window.add(sample);
+
+    }
+
+    @Override
+    public void closeCleanup() throws IOException {
+        flushResults();
     }
 
     private boolean shouldFlush(Sample sample) {
@@ -191,92 +107,4 @@ public final class BatchPipelineStep extends AbstractPipelineStep implements Tre
         }
         return false;
     }
-
-    @Override
-    protected synchronized void doClose() throws IOException {
-        if (mapMode) {
-            // Flush all remaining samples.
-            flushMapResults(-1);
-        }
-        flushResults();
-        super.doClose();
-    }
-
-    private synchronized boolean flushResults() throws IOException {
-        if (window.isEmpty())
-            return false;
-        printFlushMessage();
-        flush(window);
-        window.clear();
-        return true;
-    }
-
-    private synchronized void flushMapResults(long currentTime) throws IOException {
-        //Check for Lists to be flushed according to timeout
-        List<String> removeTags = new ArrayList<>();
-        for (String tagValue : tagList) {
-            long endTime = timeoutMap.get(tagValue);
-            if (endTime < currentTime || currentTime == -1) {
-                logger.log(Level.INFO, String.format("Flushing one result, Map Size: %s", timeoutMap.size()));
-                logger.log(Level.INFO, String.format("Flushing one result, Sample-receive: %s", samplesForTag.get(tagValue).get(0).getTag("received")));
-                removeTags.add(tagValue);
-                //Flush this window
-                for (Sample s : samplesForTag.get(tagValue)) {
-                    window.add(s);
-                }
-                flushResults();
-            } else {
-                //List 'tagList' keeps sorting of timeouts indirectly, so we can break the for loop after first not-timed-out sample
-                break;
-            }
-        }
-        for (String tagValue : removeTags) {
-            timeoutMap.remove(tagValue);
-            samplesForTag.remove(tagValue);
-            tagList.remove(tagValue);
-        }
-    }
-
-    private void printFlushMessage() {
-        String info = "";
-        if (window.isEmpty()) {
-            info = " (no samples)";
-        } else {
-            int numSamples = window.size();
-            int numMetrics = window.get(0).getMetrics().length;
-            info += "(" + numSamples + " samples, " + numMetrics + " metrics)";
-        }
-        logger.info(toString() + " computing results " + info);
-    }
-
-    private void flush(List<Sample> window) throws IOException {
-        for (BatchHandler handler : handlers) {
-            window = handler.handleBatch(window);
-        }
-        for (Sample sample : window) {
-            this.output().writeSample(sample);
-        }
-    }
-
-    // =================================================
-    // Printing ========================================
-    // =================================================
-    public String toString() {
-        StringBuilder sb = new StringBuilder();
-        for (int i = 0; i < handlers.size(); i++) {
-            sb.append(handlers.get(i).getClass().getSimpleName());
-            if (i < handlers.size() - 1) {
-                sb.append(", ");
-            }
-        }
-        return String.format("Batch processing, %s handler(s): %s", handlers.size(), sb.toString());
-    }
-
-    @Override
-    public Collection<Object> formattedChildren() {
-        List<Object> children = new ArrayList<>(handlers.size());
-        children.addAll(handlers);
-        return children;
-    }
-
 }
